+++ pdk8s/cli.py
@@ -10,9 +10,9 @@ import pdk8s.synth
 from . import k8s
 
 
-option_output = click.option("--output", default="dist", type=click.Path(file_okay=False), help="Folder in which to store generated chart")
+option_input = click.option("-i", "--input", "input_path", default="chart.py", type=click.Path(exists=True, dir_okay=False), help="Python file that generates your chart")
+option_output = click.option("-o", "--output", "output_path", default="dist", type=click.Path(file_okay=False), help="Folder in which to store generated chart")
 option_format = click.option("--format", "output_format", default="helm", type=click.Choice(['helm', 'kubernetes'], case_sensitive=False), help="Format of output")
-option_input = click.option("--input", default="chart.py", type=click.Path(exists=True, dir_okay=False), help="Python file that generates your chart")
 
 
 @click.group()
@@ -32,13 +32,13 @@ def add_subcommands(main, with_input):
     @option_output
     @option_format
     @click.pass_context
-    def synth(ctx, output, output_format, input_path=None):
+    def synth(ctx, output_path, output_format, input_path=None):
         if input_path:
             chart_variables = load_chart_from_file(input_path)
         else:
             chart_variables = ctx.obj
 
-        return pdk8s.synth.synth(chart_variables, output_format, output)
+        return pdk8s.synth.synth(chart_variables, output_format, output_path)
 
     if with_input:
         option_input(synth)